+++ alicompat/forge/src/compat/gtceu/java/com/yanny/alicompat/compat/gtceu/GregTechCEuModernLang.java
@@ -4,6 +4,7 @@ import com.yanny.aci.language.CoreLang;
 import com.yanny.aci.language.ITooltipKey;
 import com.yanny.aci.language.Translation;
 import com.yanny.alicompat.ICompatTranslations;
+import com.yanny.alicompat.Utils;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
@@ -21,7 +22,7 @@ public class GregTechCEuModernLang implements ICompatTranslations {
         private final Translation translation;
 
         Functions(String k, String e) {
-            translation = new Translation("alicompat.type.function." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "type.function", k), e);
         }
 
         @NotNull