+++ alicompat/forge/src/compat/hybridaquatic/java/com/yanny/alicompat/compat/hybridaquatic/HybridAquaticLang.java
@@ -1,28 +1,45 @@
-package com.yanny.alicompat.compat.sawmill;
+package com.yanny.alicompat.compat.hybridaquatic;
 
 import com.yanny.aci.language.CoreLang;
 import com.yanny.aci.language.ITooltipKey;
 import com.yanny.aci.language.Translation;
 import com.yanny.alicompat.ICompatTranslations;
+import com.yanny.alicompat.Utils;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class SawmillLang implements ICompatTranslations {
-    static final String MOD_ID = "sawmill";
+public class HybridAquaticLang implements ICompatTranslations {
+    static final String MOD_ID = "hybrid_aquatic";
 
     public static final Map<String, String> TRANSLATION_MAP = new HashMap<>();
 
+    public enum Entry implements ITooltipKey {
+        MESSAGE_IN_A_BOTTLE("message_in_a_bottle", "Message In A Bottle:"),
+        ;
+
+        private final Translation translation;
+
+        Entry(String k, String e) {
+            translation = new Translation(Utils.langKey(MOD_ID, "type.entry", k), e);
+        }
+
+        @NotNull
+        @Override
+        public Translation getTranslation() {
+            return translation;
+        }
+    }
+
     public enum Value implements ITooltipKey {
-        ANY_WOOD_TYPE("any_wood_type", "Any wood type"),
-        VILLAGE_WOOD_TYPE("village_wood_type", "Wood type of the village"),
+        RANDOM_SEA_MESSAGE("random_sea_message", "Random sea message"),
         ;
 
         private final Translation translation;
 
         Value(String k, String e) {
-            translation = new Translation("alicompat.property.value." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "property.value", k), e);
         }
 
         @NotNull
@@ -33,6 +50,7 @@ public class SawmillLang implements ICompatTranslations {
     }
 
     static {
+        CoreLang.register(TRANSLATION_MAP, Entry.class);
         CoreLang.register(TRANSLATION_MAP, Value.class);
     }