+++ alicompat/neoforge/src/compat/artifacts/java/com/yanny/alicompat/compat/artifacts/ArtifactsLang.java
@@ -4,6 +4,7 @@ import com.yanny.aci.language.CoreLang;
 import com.yanny.aci.language.ITooltipKey;
 import com.yanny.aci.language.Translation;
 import com.yanny.alicompat.ICompatTranslations;
+import com.yanny.alicompat.Utils;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
@@ -23,7 +24,7 @@ public class ArtifactsLang implements ICompatTranslations {
         private final Translation translation;
 
         Conditions(String k, String e) {
-            translation = new Translation("alicompat.type.condition." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "type.condition", k), e);
         }
 
         @NotNull
@@ -41,7 +42,7 @@ public class ArtifactsLang implements ICompatTranslations {
         private final Translation translation;
 
         ConsumeEffects(String k, String e) {
-            translation = new Translation("alicompat.type.consume_effect." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "type.consume_effect", k), e);
         }
 
         @NotNull
@@ -58,7 +59,7 @@ public class ArtifactsLang implements ICompatTranslations {
         private final Translation translation;
 
         Functions(String k, String e) {
-            translation = new Translation("alicompat.type.function." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "type.function", k), e);
         }
 
         @NotNull
@@ -76,7 +77,7 @@ public class ArtifactsLang implements ICompatTranslations {
         private final Translation translation;
 
         Value(String k, String e) {
-            translation = new Translation("alicompat.property.value." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "property.value", k), e);
         }
 
         @NotNull