+++ alicompat/neoforge/src/compat/immersiveengineering/java/com/yanny/alicompat/compat/immersiveengineering/ImmersiveEngineeringLang.java
@@ -4,6 +4,7 @@ import com.yanny.aci.language.CoreLang;
 import com.yanny.aci.language.ITooltipKey;
 import com.yanny.aci.language.Translation;
 import com.yanny.alicompat.ICompatTranslations;
+import com.yanny.alicompat.Utils;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
@@ -21,7 +22,7 @@ public class ImmersiveEngineeringLang implements ICompatTranslations {
         private final Translation translation;
 
         Conditions(String k, String e) {
-            translation = new Translation("alicompat.type.condition." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "type.condition", k), e);
         }
 
         @NotNull
@@ -40,7 +41,7 @@ public class ImmersiveEngineeringLang implements ICompatTranslations {
         private final Translation translation;
 
         Entry(String k, String e) {
-            translation = new Translation("alicompat.type.entry." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "type.entry", k), e);
         }
 
         @NotNull
@@ -61,7 +62,7 @@ public class ImmersiveEngineeringLang implements ICompatTranslations {
         private final Translation translation;
 
         Functions(String k, String e) {
-            translation = new Translation("alicompat.type.function." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "type.function", k), e);
         }
 
         @NotNull
@@ -78,7 +79,7 @@ public class ImmersiveEngineeringLang implements ICompatTranslations {
         private final Translation translation;
 
         ItemSubPredicates(String k, String e) {
-            translation = new Translation("alicompat.type.item_sub_predicate." + k, e);
+            translation = new Translation(Utils.langKey(MOD_ID, "type.item_sub_predicate", k), e);
         }
 
         @NotNull
@@ -95,7 +96,7 @@ public class ImmersiveEngineeringLang implements ICompatTranslations {
         private final Translation translation;
 
         Value(String k, String e) {
-            this.translation = new Translation("alicompat.property.value." + k, e);
+            this.translation = new Translation(Utils.langKey(MOD_ID, "property.value", k), e);
         }
 
         @NotNull
@@ -112,7 +113,7 @@ public class ImmersiveEngineeringLang implements ICompatTranslations {
         private final Translation translation;
 
         Branch(ITooltipKey s, String k, String e) {
-            this.translation = new Translation(s.singular(), "alicompat.property.branch." + k, s.englishSingular(), e);
+            this.translation = new Translation(s.singular(), Utils.langKey(MOD_ID, "property.branch", k), s.englishSingular(), e);
         }
 
         @NotNull