+++ alicompat/neoforge/src/compat/occultism/java/com/yanny/alicompat/compat/occultism/AddItemModifierAccessor.java
@@ -35,7 +35,7 @@ public class AddItemModifierAccessor extends BaseAccessor<AddItemModifier> imple
         List<LootItemCondition> conditionList = NodeUtils.unwrapConditions(utils, condition);
 
         return Optional.of(GlobalLootModifierUtils.getLootModifier(utils, parent, conditionList,
-                (c) -> Collections.singletonList(new IOperation.AddOperation((itemStack) -> true,
+                (page, c) -> Collections.singletonList(new IOperation.AddOperation((itemStack) -> true,
                         GlmNodeUtils.addedNode(utils, c, addedItem.getDefaultInstance(), 1, new RangeValue(count))))));
     }
 }