+++ alicompat/CHANGELOG.md
@@ -1,9 +1,8 @@
 ## []
 
+- Added Farmer's Delight support
 - Added Occultism support
 - Added Repurposed Structures support
-- Added Sophisticated Backpacks support
-- Added Sophisticated Storage support
 
 ## [1.0.0]